refactor: consolidated sidecar examples
