--- nextjs-sidecar-with-hooks/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "nextjs-sidecar-with-hooks",
-    "version": "0.1.0",
-    "private": true,
-    "author": {
-        "name": "Matthew Kinbaum",
-        "email": "mkinbaum@gmail.com"
-    },
-    "scripts": {
-        "dev": "next dev -p 8080",
-        "build": "next build",
-        "start": "next start -p 8080"
-    },
-    "dependencies": {
-        "next": "9.4.4",
-        "react": "16.13.1",
-        "react-dom": "16.13.1"
-    },
-    "devDependencies": {
-        "babel-eslint": "^10.1.0",
-        "eslint": "^7.4.0",
-        "eslint-config-airbnb": "^18.2.0",
-        "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-jsx-a11y": "^6.3.1",
-        "eslint-plugin-react": "^7.20.3",
-        "eslint-plugin-react-hooks": "^4.0.5"
-    }
-}
