--- nextjs-sidecar-with-hooks/sidecar/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "sidecar",
-    "version": "1.0.0",
-    "license": "MIT",
-    "scripts": {
-        "build": "webpack --mode development",
-        "start": "cd dist && PORT=8081 npx servor"
-    },
-    "author": {
-        "name": "Matthew Kinbaum",
-        "email": "mkinbaum@gmail.com"
-    },
-    "dependencies": {
-        "react": "^16.13.1"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.10.3",
-        "@babel/preset-env": "^7.10.3",
-        "@babel/preset-react": "^7.10.1",
-        "babel-loader": "^8.1.0",
-        "webpack": "^5.0.0-beta.20",
-        "webpack-cli": "^3.3.12"
-    }
-}
