--- nextjs-sidecar-with-hooks/sidecar/src/index.js
@@ -1,1 +1,0 @@
-export default () => true;
