+++ nextjs-sidecar/dog-admin/package.json
@@ -9,8 +9,8 @@
   },
   "license": "MIT",
   "author": {
-    "name": "Jack Herrington",
-    "email": "jherr@pobox.com"
+    "name": "Matthew Kinbaum",
+    "email": "mkinbaum@gmail.com"
   },
   "devDependencies": {
     "@babel/core": "7.10.4",
@@ -20,12 +20,15 @@
     "css-loader": "3.6.0",
     "html-webpack-plugin": "git://github.com/ScriptedAlchemy/html-webpack-plugin#master",
     "style-loader": "1.2.1",
-    "webpack": "5.0.0-beta.18",
+    "webpack": "5.0.0-beta.20",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.11.0"
   },
   "dependencies": {
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0"
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  },
+  "resolutions": {
+    "webpack": "5.0.0-beta.20"
   }
 }
