+++ nextjs-sidecar/dog-admin/src/DogName.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function DogName({ name }) {
+  return <h1>Look at these {name}</h1>;
+}
