--- nextjs-sidecar/dog-admin/src/DogName.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export default ({ name }) => (
-  <h1
-    style={{
-      color: "red",
-    }}
-  >
-    Dog's name is {name}
-  </h1>
-);
