--- nextjs-sidecar/host/components/Dog.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-export default () => (
-  <img
-    src="https://placedog.net/500/280?id=65"
-    style={{
-      padding: "1em",
-      border: "10px solid green",
-      width: 500,
-      height: 280,
-    }}
-  />
-);
