+++ nextjs-sidecar/host/package.json
@@ -2,6 +2,10 @@
   "name": "@nextjs-sidecar/host",
   "version": "0.1.0",
   "private": true,
+  "author": {
+    "name": "Matthew Kinbaum",
+    "email": "mkinbaum@gmail.com"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -11,5 +15,13 @@
     "next": "9.4.4",
     "react": "16.13.1",
     "react-dom": "16.13.1"
+  },
+  "devDependencies": {
+    "eslint": "^7.4.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-react-hooks": "^4.0.5"
   }
 }
