+++ nextjs-sidecar/host/sidecar/.gitignore
