+++ nextjs-sidecar/host/sidecar/package.json
@@ -7,18 +7,21 @@
   },
   "license": "MIT",
   "author": {
-    "name": "Jack Herrington",
-    "email": "jherr@pobox.com"
+    "name": "Matthew Kinbaum",
+    "email": "mkinbaum@gmail.com"
   },
   "devDependencies": {
     "@babel/core": "7.10.4",
     "@babel/preset-env": "7.10.4",
     "@babel/preset-react": "7.10.4",
     "babel-loader": "8.1.0",
-    "webpack": "5.0.0-beta.18",
+    "webpack": "5.0.0-beta.20",
     "webpack-cli": "3.3.11"
   },
   "dependencies": {
-    "react": "^16.12.0"
+    "react": "16.13.1"
+  },
+  "resolutions": {
+    "webpack": "5.0.0-beta.20"
   }
 }
