+++ nextjs-sidecar/host/sidecar/webpack.config.js
@@ -1,39 +1,45 @@
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
+const { ModuleFederationPlugin } = require("webpack").container;
+const { dependencies } = require("../package.json");
 
 module.exports = {
   output: {
     publicPath: "http://localhost:8080/",
   },
-
   resolve: {
     extensions: [".jsx", ".js"],
   },
-
+  devServer: {
+    port: 8081,
+  },
   module: {
     rules: [
       {
-        test: /\.(js|jsx)$/,
+        test: /\.js$/,
         exclude: /node_modules/,
         use: {
           loader: "babel-loader",
           options: {
-            presets: ["@babel/preset-react"],
+            presets: ["@babel/preset-react", "@babel/preset-env"],
           },
         },
       },
     ],
   },
-
   plugins: [
     new ModuleFederationPlugin({
       name: "dogs",
       library: { type: "var", name: "dogs" },
       filename: "remoteEntry.js",
-      remotes: {},
       exposes: {
-        "./Dog": "../components/Dog",
+        "./Dog": "../src/components/Dog",
       },
-      shared: ["react"],
+      shared: {
+        ...dependencies,
+        react: {
+          // import: false,
+          singleton: true,
+        },
+      },
     }),
   ],
 };
